feat: use tippy
