+++ src/components/Publication/QueuedPublication.tsx
@@ -91,7 +91,7 @@ const QueuedPublication: FC<Props> = ({ txn }) => {
   return (
     <article className="p-5">
       <div className="pb-4 flex items-start justify-between">
-        <UserProfile profile={currentProfile as Profile} showPreviewCard />
+        <UserProfile profile={currentProfile as Profile} />
         <Tooltip content="Indexing" placement="top">
           <div className="bg-brand-200 rounded-full h-4 w-4 flex items-center justify-center">
             <div className="animate-pulse bg-brand-500 rounded-full h-2 w-2" />
