+++ src/components/Publication/SinglePublication.tsx
@@ -58,7 +58,7 @@ const SinglePublication: FC<Props> = ({
       )}
       <div className="flex justify-between pb-4 space-x-1.5">
         <span onClick={(event) => event.stopPropagation()}>
-          <UserProfile profile={profile ?? publication?.collectedBy?.defaultProfile} showPreviewCard />
+          <UserProfile profile={profile ?? publication?.collectedBy?.defaultProfile} />
         </span>
         <span className="text-xs text-gray-500">{dayjs(new Date(timestamp)).fromNow()}</span>
       </div>
