+++ src/components/Publication/index.tsx
@@ -85,7 +85,6 @@ const ViewPublication: NextPage = () => {
               publication.__typename === 'Mirror' ? publication?.mirrorOf?.profile : publication?.profile
             }
             showBio
-            showPreviewCard
           />
         </Card>
         <RelevantPeople publication={publication} />
