+++ src/components/Shared/UserPreview.tsx
@@ -1,0 +1,161 @@
+import MutualFollowers from '@components/Profile/MutualFollowers';
+import type { Profile } from '@generated/types';
+import { BadgeCheckIcon } from '@heroicons/react/solid';
+import getAvatar from '@lib/getAvatar';
+import isVerified from '@lib/isVerified';
+import Tippy from '@tippyjs/react';
+import clsx from 'clsx';
+import type { Dispatch, FC } from 'react';
+import { useEffect } from 'react';
+import { useState } from 'react';
+import React, { Fragment, useRef } from 'react';
+import { useAppStore } from 'src/store/app';
+
+import Follow from './Follow';
+import Markup from './Markup';
+import Slug from './Slug';
+import SuperFollow from './SuperFollow';
+
+type Props = {
+  isBig: boolean;
+  profile: Profile;
+  followStatusLoading: boolean;
+  showBio: boolean;
+  following: boolean;
+  setFollowing: Dispatch<boolean>;
+};
+
+const UserPreview: FC<Props> = ({
+  profile,
+  isBig,
+  followStatusLoading,
+  showBio,
+  following,
+  setFollowing
+}) => {
+  const wrapperRef = useRef<HTMLDivElement>(null);
+  const [showPreview, setShowPreview] = useState(false);
+  const currentProfile = useAppStore((state) => state.currentProfile);
+
+  const UserImage = () => (
+    <img
+      src={getAvatar(profile)}
+      loading="lazy"
+      className={clsx(
+        isBig ? 'w-14 h-14' : 'w-10 h-10',
+        'bg-gray-200 rounded-full border dark:border-gray-700/80'
+      )}
+      height={isBig ? 56 : 40}
+      width={isBig ? 56 : 40}
+      alt={profile?.handle}
+    />
+  );
+
+  const UserName = () => (
+    <>
+      <div className="flex gap-1 items-center max-w-sm truncate">
+        <div className={clsx(isBig ? 'font-bold' : 'text-md')}>{profile?.name ?? profile?.handle}</div>
+        {isVerified(profile?.id) && <BadgeCheckIcon className="w-4 h-4 text-brand" />}
+      </div>
+      <Slug className="text-sm" slug={profile?.handle} prefix="@" />
+    </>
+  );
+
+  const Preview = () => (
+    <>
+      <div className="flex justify-between items-center">
+        <UserImage />
+        <div onClick={(e) => e.preventDefault()}>
+          {!profile.isFollowedByMe &&
+            (followStatusLoading ? (
+              <div className="w-10 h-8 rounded-lg shimmer" />
+            ) : following ? null : profile?.followModule?.__typename === 'FeeFollowModuleSettings' ? (
+              <SuperFollow profile={profile} setFollowing={setFollowing} />
+            ) : (
+              <Follow profile={profile} setFollowing={setFollowing} />
+            ))}
+        </div>
+      </div>
+      <div className="p-1 space-y-3">
+        <UserName />
+        <div>
+          {profile?.bio && (
+            <div className={clsx(isBig ? 'text-base' : 'text-sm', 'mt-2', 'linkify leading-6')}>
+              <Markup>{profile?.bio}</Markup>
+            </div>
+          )}
+        </div>
+        <div className="flex space-x-3 items-center">
+          <div className="flex items-center space-x-1">
+            <div className="text-base">{profile?.stats?.totalFollowing}</div>
+            <div className="text-gray-500 text-sm">Following</div>
+          </div>
+          <div className="flex items-center space-x-1 text-md">
+            <div className="text-base">{profile?.stats?.totalFollowers}</div>
+            <div className="text-gray-500 text-sm">Followers</div>
+          </div>
+        </div>
+        {currentProfile && <MutualFollowers profile={profile} variant="xs" />}
+      </div>
+    </>
+  );
+
+  const Trigger = () => (
+    <div className="relative flex items-center space-x-3">
+      <UserImage />
+      <div>
+        <UserName />
+        {showBio && profile?.bio && (
+          <div className={clsx(isBig ? 'text-base' : 'text-sm', 'mt-2', 'linkify leading-6')}>
+            <Markup>{profile?.bio}</Markup>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+
+  const onPreviewEnd = () => {
+    setShowPreview(false);
+  };
+
+  const onPreviewStart = () => {
+    setShowPreview(true);
+  };
+
+  useEffect(() => {
+    const handleHoverOutside = (event: any) => {
+      if (wrapperRef.current && !wrapperRef.current?.contains(event.target)) {
+        onPreviewEnd();
+      }
+    };
+    document.addEventListener('mouseleave', handleHoverOutside);
+    return () => {
+      document.removeEventListener('mouseleave', handleHoverOutside);
+    };
+  }, [wrapperRef]);
+
+  return (
+    <div onMouseOver={onPreviewStart} ref={wrapperRef}>
+      {showPreview ? (
+        <Tippy
+          placement="bottom-start"
+          delay={500}
+          hideOnClick={false}
+          content={<Preview />}
+          arrow={false}
+          interactive
+          zIndex={1000}
+          className="!bg-white !-my-2 !text-black !px-1.5 !py-3 dark:!text-white !w-64 dark:!bg-black !border dark:!border-gray-700 !rounded-xl"
+        >
+          <span>
+            <Trigger />
+          </span>
+        </Tippy>
+      ) : (
+        <Trigger />
+      )}
+    </div>
+  );
+};
+
+export default UserPreview;
